bolt-intro-python/requirements.txt
autopep8==1.6.0
pycodestyle==2.8.0
python-dotenv==0.19.2
slack-bolt==1.10.0
slack-sdk==3.12.0
toml==0.10.2
